pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "resume-optimizer"
version = "0.1.0"
description = "AI-powered resume optimization tool with ATS compatibility checking"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["resume", "optimization", "ats", "ai", "nlp", "streamlit"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Text Processing",
]

# Core dependencies
dependencies = [
    # Web framework
    "streamlit>=1.28.0",
    # AI and LangChain
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-perplexity>=0.1.0",
    "langchain-google-genai>=1.0.0",
    # NLP and text processing
    "spacy>=3.7.0",
    "nltk>=3.8.0",
    # File processing
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    # PDF generation
    "reportlab>=4.0.0",
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
    # Data science
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    # HTTP and API clients
    "requests>=2.31.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    # Configuration and utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "pip>=25.2",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/resume-optimizer"
Repository = "https://github.com/yourusername/resume-optimizer.git"
Documentation = "https://github.com/yourusername/resume-optimizer/blob/main/README.md"
"Bug Reports" = "https://github.com/yourusername/resume-optimizer/issues"

[project.scripts]
resume_optimizer = "resume_optimizer.main:main"
resume-optimizer = "resume_optimizer.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
    "ruff>=0.13.2",
]

[tool.hatchling.build.targets.wheel]
packages = ["src/resume_optimizer"]

# Tool configurations
[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["resume_optimizer"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503", "E501"]
exclude = [".git", ".venv", "build", "dist", "__pycache__"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"
